+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Refactor projects on nav bar to be a tree
 - Edited appearnce of settings page to include checkboxes and configuration details
 - Refactor project status/health into a Radix UI `<Card>` and `<HoverCard>`
+- creating project will redirect to auto submitted goals page instead of to the projects page
 
 ### Bug Fixes